my_agent/requirements.txt
# Core dependencies for SQL Query Writer Agent
# Add your additional packages below with pinned versions

# Database
duckdb==1.1.3
sqlalchemy==2.0.36

# Data loading
kagglehub==0.3.4

# Ollama (required - for LLM access)
ollama==0.4.7

# Machine learning — few-shot retrieval (TF-IDF cosine similarity)
scikit-learn==1.6.0
numpy==2.2.0

# Environment variable loading (.env support)
python-dotenv==1.0.1

# GCLS beta-curve plot (Step 6, optional --plot flag)
matplotlib==3.10.8
